input and radio button components refactored
