+++ src/core/InputField/InputField.stories.tsx
@@ -1,8 +1,8 @@
+import { AccountCircle } from '@material-ui/icons';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
 import { ThemedBackground } from '../../utils/storybook';
 import { InputField } from '../InputField';
-import { AccountCircle } from '@material-ui/icons';
 storiesOf('InputField', module)
   // Litmus Portal
   .add('Litmus Portal', () => (
@@ -11,46 +11,25 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="error"
           variant="error"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
-        <InputField
-          label="disabled"
-          variant="primary"
-          type="text"
-          disabled={true}
-          endIcon={null}
-          startIcon={null}
-        />
+        <InputField label="disabled" variant="primary" disabled={true} />
       </div>
       <div style={{ display: 'grid', gridGap: '20px' }}>
         <InputField
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
-          startIcon={null}
           value="Input with password type"
           onChange={(e) => e.target.value}
         />
@@ -58,7 +37,6 @@ storiesOf('InputField', module)
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
           startIcon={<AccountCircle />}
           value="Input with password type and start icon"
           onChange={(e) => e.target.value}
@@ -66,29 +44,21 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with Start adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
-          endIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with End adornment"
-          disabled={false}
-          startIcon={null}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with both adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
@@ -96,30 +66,19 @@ storiesOf('InputField', module)
         <InputField
           label="error"
           variant="error"
-          type="text"
           value="Input with error type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
           value="Input with success type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="disabled"
           variant="primary"
-          type="text"
           disabled={true}
-          endIcon={null}
-          startIcon={null}
           value="disabled text"
         />
       </div>
@@ -133,46 +92,25 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="error"
           variant="error"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
-        <InputField
-          label="disabled"
-          variant="primary"
-          type="text"
-          disabled={true}
-          endIcon={null}
-          startIcon={null}
-        />
+        <InputField label="disabled" variant="primary" disabled={true} />
       </div>
       <div style={{ display: 'grid', gridGap: '20px' }}>
         <InputField
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
-          startIcon={null}
           value="Input with password type"
           onChange={(e) => e.target.value}
         />
@@ -180,7 +118,6 @@ storiesOf('InputField', module)
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
           startIcon={<AccountCircle />}
           value="Input with password type"
           onChange={(e) => e.target.value}
@@ -188,29 +125,21 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with Start adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
-          endIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with End adornment"
-          disabled={false}
-          startIcon={null}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with both adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
@@ -218,30 +147,19 @@ storiesOf('InputField', module)
         <InputField
           label="error"
           variant="error"
-          type="text"
           value="Input with error type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
           value="Input with success type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="disabled"
           variant="primary"
-          type="text"
           disabled={true}
-          endIcon={null}
-          startIcon={null}
           value="disabled text"
         />
       </div>
@@ -255,46 +173,25 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="error"
           variant="error"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
-        <InputField
-          label="disabled"
-          variant="primary"
-          type="text"
-          disabled={true}
-          endIcon={null}
-          startIcon={null}
-        />
+        <InputField label="disabled" variant="primary" disabled={true} />
       </div>
       <div style={{ display: 'grid', gridGap: '20px' }}>
         <InputField
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
-          startIcon={null}
           value="Input with password type"
           onChange={(e) => e.target.value}
         />
@@ -302,7 +199,6 @@ storiesOf('InputField', module)
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
           startIcon={<AccountCircle />}
           value="Input with password type"
           onChange={(e) => e.target.value}
@@ -310,29 +206,21 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with Start adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
-          endIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with End adornment"
-          disabled={false}
-          startIcon={null}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with both adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
@@ -340,30 +228,19 @@ storiesOf('InputField', module)
         <InputField
           label="error"
           variant="error"
-          type="text"
           value="Input with error type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
           value="Input with success type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="disabled"
           variant="primary"
-          type="text"
           disabled={true}
-          endIcon={null}
-          startIcon={null}
           value="disabled text"
         />
       </div>
@@ -377,46 +254,25 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="error"
           variant="error"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
-        <InputField
-          label="disabled"
-          variant="primary"
-          type="text"
-          disabled={true}
-          endIcon={null}
-          startIcon={null}
-        />
+        <InputField label="disabled" variant="primary" disabled={true} />
       </div>
       <div style={{ display: 'grid', gridGap: '20px' }}>
         <InputField
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
-          startIcon={null}
           value="Input with password type"
           onChange={(e) => e.target.value}
         />
@@ -424,7 +280,6 @@ storiesOf('InputField', module)
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
           startIcon={<AccountCircle />}
           value="Input with password type"
           onChange={(e) => e.target.value}
@@ -432,29 +287,21 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with Start adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
-          endIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with End adornment"
-          disabled={false}
-          startIcon={null}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with both adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
@@ -462,30 +309,19 @@ storiesOf('InputField', module)
         <InputField
           label="error"
           variant="error"
-          type="text"
           value="Input with error type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
           value="Input with success type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="disabled"
           variant="primary"
-          type="text"
           disabled={true}
-          endIcon={null}
-          startIcon={null}
           value="disabled text"
         />
       </div>
@@ -499,46 +335,25 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="error"
           variant="error"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
-        <InputField
-          label="disabled"
-          variant="primary"
-          type="text"
-          disabled={true}
-          endIcon={null}
-          startIcon={null}
-        />
+        <InputField label="disabled" variant="primary" disabled={true} />
       </div>
       <div style={{ display: 'grid', gridGap: '20px' }}>
         <InputField
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
-          startIcon={null}
           value="Input with password type"
           onChange={(e) => e.target.value}
         />
@@ -546,7 +361,6 @@ storiesOf('InputField', module)
           label="primary"
           variant="primary"
           type="password"
-          disabled={false}
           startIcon={<AccountCircle />}
           value="Input with password type"
           onChange={(e) => e.target.value}
@@ -554,29 +368,21 @@ storiesOf('InputField', module)
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with Start adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
-          endIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with End adornment"
-          disabled={false}
-          startIcon={null}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="primary"
           variant="primary"
-          type="text"
           value="Input with both adornment"
-          disabled={false}
           startIcon={<AccountCircle />}
           endIcon={<AccountCircle />}
           onChange={(e) => e.target.value}
@@ -584,30 +390,19 @@ storiesOf('InputField', module)
         <InputField
           label="error"
           variant="error"
-          type="text"
           value="Input with error type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="success"
           variant="success"
-          type="text"
           value="Input with success type"
-          disabled={false}
-          endIcon={null}
-          startIcon={null}
           onChange={(e) => e.target.value}
         />
         <InputField
           label="disabled"
           variant="primary"
-          type="text"
           disabled={true}
-          endIcon={null}
-          startIcon={null}
           value="disabled text"
         />
       </div>
